split up remove_freeze
