+++ candy-machine/program/src/lib.rs
@@ -13,7 +13,6 @@ declare_id!("cndy3Z4yapfJBmL3ShUp5exZKqR3z33thTzeNMm2gRZ");
 
 #[program]
 pub mod candy_machine {
-
     use super::*;
 
     pub fn initialize_candy_machine(
@@ -74,13 +73,15 @@ pub mod candy_machine {
         handle_set_freeze(ctx, freeze_time)
     }
 
-    pub fn remove_freeze<'info>(
-        ctx: Context<'_, '_, '_, 'info, RemoveFreeze<'info>>,
-    ) -> Result<()> {
+    pub fn remove_freeze(ctx: Context<RemoveFreeze>) -> Result<()> {
         handle_remove_freeze(ctx)
     }
 
     pub fn thaw_nft(ctx: Context<ThawNFT>) -> Result<()> {
         handle_thaw_nft(ctx)
+    }
+
+    pub fn unlock_funds<'info>(ctx: Context<'_, '_, '_, 'info, UnlockFunds<'info>>) -> Result<()> {
+        handle_unlock_funds(ctx)
     }
 }
