+++ candy-machine/program/src/processor/freeze/mod.rs
@@ -1,7 +1,9 @@
 pub mod remove_freeze;
 pub mod set_freeze;
 pub mod thaw_nft;
+pub mod unlock_funds;
 
 pub use remove_freeze::*;
 pub use set_freeze::*;
 pub use thaw_nft::*;
+pub use unlock_funds::*;
